♻️ refactor: PvP 컨텐츠 파일 폴더화
