+++ src/handler/pvp/states/result/pvpGameOverState.js
@@ -1,13 +1,13 @@
 // src/handler/pvp/states/pvpGameOverState.js
 
 import sessionManager from '#managers/sessionManager.js';
-import { PacketType } from '../../../constants/header.js';
-import { createResponse } from '../../../utils/response/createResponse.js';
-import { PVP_STATUS } from '../../../constants/battle.js';
-import PvpState from './pvpState.js';
-import { getPlayerRatingFromRedis, updatePlayerRating } from '../../../db/redis/ratingService.js';
-import { invalidResponseCode } from '../../../utils/error/invalidResponseCode.js';
-import logger from '../../../utils/log/logger.js';
+import { PacketType } from '../../../../constants/header.js';
+import { createResponse } from '../../../../utils/response/createResponse.js';
+import { PVP_STATUS } from '../../../../constants/battle.js';
+import PvpState from '../base/pvpState.js';
+import { getPlayerRatingFromRedis, updatePlayerRating } from '../../../../db/redis/ratingService.js';
+import { invalidResponseCode } from '../../../../utils/error/invalidResponseCode.js';
+import logger from '../../../../utils/log/logger.js';
 
 const RANK_CHANGE_POINTS = 10;
 
@@ -15,7 +15,6 @@ export default class PvpGameOverState extends PvpState {
   async enter() {
     this.pvpRoom.pvpStatus = PVP_STATUS.GAME_OVER;
 
-    // 유저 버프 초기화
     this.mover.buff = null;
     this.mover.battleCry = false;
     this.mover.berserk = false;
